post de trayectos y entidades asociadas

+++ pkg/entities/stop.go
@@ -4,7 +4,7 @@ import "gorm.io/gorm"
 
 type Stop struct {
 	gorm.Model
-	Orden       uint
+	Orden       int64
 	TrayectosID uint
 	AddressesID uint
 	Trayecto    Trayecto `gorm:"foreignKey:TrayectosID"`
